study groups and course learning resource

+++ MainService/src/main/java/ru/espada/ep/iptip/course/CourseService.java
@@ -1,15 +1,20 @@
 package ru.espada.ep.iptip.course;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.espada.ep.iptip.course.model.CourseResponseModel;
+import ru.espada.ep.iptip.course.learning_resource_category.CourseLearningResourceCategoryEntity;
+import ru.espada.ep.iptip.course.learning_resource_category.CourseLearningResourceCategoryRepository;
 import ru.espada.ep.iptip.course.model.CreateCourseModel;
-import ru.espada.ep.iptip.course.test.TestEntity;
+import ru.espada.ep.iptip.course.model.CreateCourseLearningResourceCategoryModel;
+import ru.espada.ep.iptip.course.model.StudyGroupCourseModel;
 import ru.espada.ep.iptip.course.user_course.UserCourseEntity;
 import ru.espada.ep.iptip.course.user_course.UserCourseRepository;
 import ru.espada.ep.iptip.study_groups.StudyGroupEntity;
 import ru.espada.ep.iptip.study_groups.StudyGroupRepository;
+import ru.espada.ep.iptip.study_groups.StudyGroupService;
+import ru.espada.ep.iptip.university.institute.major.faculty.FacultyEntity;
 import ru.espada.ep.iptip.user.UserEntity;
 import ru.espada.ep.iptip.user.UserRepository;
 import ru.espada.ep.iptip.user.UserService;
@@ -17,6 +22,7 @@ import ru.espada.ep.iptip.user.UserService;
 import java.security.Principal;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Set;
 
 @Service
 public class CourseService {
@@ -26,18 +32,21 @@ public class CourseService {
     private StudyGroupRepository studyGroupRepository;
     private UserRepository userRepository;
     private UserService userService;
+    private StudyGroupService studyGroupService;
+    private CourseLearningResourceCategoryRepository courseLearningResourceCategoryRepository;
 
 
-    public void createCourse(Principal principal, CreateCourseModel createCourseModel) {
-        UserEntity userEntity = userRepository.findByUsername(principal.getName()).orElseThrow(() -> new RuntimeException("User not found"));
+    public Long createCourse(CreateCourseModel createCourseModel) {
         CourseEntity courseEntity = CourseEntity.builder()
                 .name(createCourseModel.getName())
                 .description(createCourseModel.getDescription())
+                .faculties(Set.of(FacultyEntity.builder().id(createCourseModel.getFacultyId()).build()))
                 .build();
         courseRepository.save(courseEntity);
+        return courseEntity.getId();
     }
 
-    public void deleteCourse(Long id) {
+    public void deleteCourse(Principal principal, Long id) {
         courseRepository.deleteById(id);
     }
 
@@ -57,19 +66,20 @@ public class CourseService {
     }
 
     @Transactional
-    public Boolean detachUserFromCourse(Long courseId, Long userId) {
+    public void detachUserFromCourse(Long courseId, Long userId) {
         UserEntity userEntity = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
         CourseEntity courseEntity = courseRepository.findById(courseId).orElseThrow(() -> new RuntimeException("Course not found"));
 
         UserCourseEntity userCourseEntity = userCourseRepository.findUserCourseEntityByUserIdAndCourseId(userEntity.getId(), courseEntity.getId());
         userCourseRepository.delete(userCourseEntity);
-        return true;
     }
 
     @Transactional
-    public Boolean attachStudyGroupToCourse(Long courseId, Long studyGroupId) {
-        CourseEntity courseEntity = courseRepository.findById(courseId).orElseThrow(() -> new RuntimeException("Course not found"));
-        StudyGroupEntity studyGroupEntity = studyGroupRepository.findById(studyGroupId).orElseThrow(() -> new RuntimeException("StudyGroup not found"));
+    public void attachStudyGroupToCourse(Principal principal, StudyGroupCourseModel attachStudyGroupToCourseModel) {
+        if (!studyGroupService.hasPermission(principal.getName(), attachStudyGroupToCourseModel.getCourseId())) throw new RuntimeException("Permission denied");
+
+        CourseEntity courseEntity = courseRepository.findById(attachStudyGroupToCourseModel.getCourseId()).orElseThrow(() -> new RuntimeException("Course not found"));
+        StudyGroupEntity studyGroupEntity = studyGroupRepository.findById(attachStudyGroupToCourseModel.getStudyGroupId()).orElseThrow(() -> new RuntimeException("StudyGroup not found"));
         List<UserEntity> users = studyGroupEntity.getUsers().stream().toList();
 
         userCourseRepository.saveAll(users.stream().map(user -> UserCourseEntity.builder()
@@ -77,13 +87,14 @@ public class CourseService {
                 .userId(user.getId())
                 .build()).toList());
 
-        return true;
     }
 
     @Transactional
-    public Boolean detachStudyGroupFromCourse(Long courseId, Long studyGroupId) {
-        CourseEntity courseEntity = courseRepository.findById(courseId).orElseThrow(() -> new RuntimeException("Course not found"));
-        StudyGroupEntity studyGroupEntity = studyGroupRepository.findById(studyGroupId).orElseThrow(() -> new RuntimeException("StudyGroup not found"));
+    public void detachStudyGroupFromCourse(Principal principal, StudyGroupCourseModel detachUserFromCourseModel) {
+        if (!studyGroupService.hasPermission(principal.getName(), detachUserFromCourseModel.getCourseId())) throw new RuntimeException("Permission denied");
+
+        CourseEntity courseEntity = courseRepository.findById(detachUserFromCourseModel.getCourseId()).orElseThrow(() -> new RuntimeException("Course not found"));
+        StudyGroupEntity studyGroupEntity = studyGroupRepository.findById(detachUserFromCourseModel.getStudyGroupId()).orElseThrow(() -> new RuntimeException("StudyGroup not found"));
         List<UserEntity> users = studyGroupEntity.getUsers().stream().toList();
 
         userCourseRepository.deleteAll(users.stream().map(user -> UserCourseEntity.builder()
@@ -91,7 +102,27 @@ public class CourseService {
                 .userId(user.getId())
                 .build()).toList());
 
-        return true;
+    }
+
+    public List<CourseEntity> getUserCourses(Long userId) {
+        return this.userCourseRepository.findUserCourseEntitiesByUserId(userId)
+                .stream()
+                .map(userCourseEntity -> courseRepository.findById(userCourseEntity.getCourseId()).orElseThrow(() -> new RuntimeException("Course not found")))
+                .sorted(Comparator.comparing(CourseEntity::getName))
+                .toList();
+    }
+
+    public Long createCourseLearningResourceCategory(CreateCourseLearningResourceCategoryModel createCourseLearningResourceCategoryModel) {
+        CourseLearningResourceCategoryEntity courseLearningResourceCategoryEntity = CourseLearningResourceCategoryEntity.builder()
+                .name(createCourseLearningResourceCategoryModel.getName())
+                .course(CourseEntity.builder().id(createCourseLearningResourceCategoryModel.getCourseId()).build())
+                .build();
+
+        return courseLearningResourceCategoryRepository.save(courseLearningResourceCategoryEntity).getId();
+    }
+
+    public void deleteCourseLearningResourceCategory(Long id) {
+        courseLearningResourceCategoryRepository.deleteById(id);
     }
 
     @Autowired
@@ -119,11 +150,13 @@ public class CourseService {
         this.userService = userService;
     }
 
-    public List<CourseEntity> getUserCourses(Long userId) {
-        return this.userCourseRepository.findUserCourseEntitiesByUserId(userId)
-                .stream()
-                .map(userCourseEntity -> courseRepository.findById(userCourseEntity.getCourseId()).orElseThrow(() -> new RuntimeException("Course not found")))
-                .sorted(Comparator.comparing(CourseEntity::getName))
-                .toList();
+    @Autowired
+    public void setCourseLearningResourceCategoryRepository(CourseLearningResourceCategoryRepository courseLearningResourceCategoryRepository) {
+        this.courseLearningResourceCategoryRepository = courseLearningResourceCategoryRepository;
+    }
+
+    @Autowired
+    public void setStudyGroupService(StudyGroupService studyGroupService) {
+        this.studyGroupService = studyGroupService;
     }
 }
